refactor: move ThemeContext creation to useTheme for ESLint compliance
