+++ dataloom-frontend/src/context/useTheme.js
@@ -1,4 +1,5 @@
-import { useContext } from "react";
-import { ThemeContext } from "./ThemeContext";
+import { createContext, useContext } from "react";
+
+export const ThemeContext = createContext();
 
 export const useTheme = () => useContext(ThemeContext);
